Big changes in Bullet's update method, implementing movement and HitBox and changin Position from int to double

+++ src/main/java/PvZ/utilities/Position.java
@@ -1,5 +1,5 @@
 package PvZ.utilities;
 
-public record Position(int x, int y) {
+public record Position(double x, double y) {
     
 }
